Migration fix

+++ db/migrate/20221216124549_create_user_reads.rb
@@ -1,8 +1,8 @@
 class CreateUserReads < ActiveRecord::Migration[7.0]
   def change
     create_table :user_reads do |t|
-      t.integer :user_id
-      t.integer :node_id
+      t.bigint :user_id
+      t.bigint :node_id
 
       t.timestamps
     end
